+++ approvaltests-util/src/test/java/com/spun/util/parser/tests/VelocityUtilsTest.java
@@ -1,12 +1,12 @@
 package com.spun.util.parser.tests;
 
 import com.spun.util.DateUtils;
+//import org.approvaltests.utils.WithTimeZone;
 import org.approvaltests.velocity.VelocityApprovals;
 import org.junit.Test;
 
 import java.sql.Timestamp;
 import java.util.Arrays;
-import java.util.TimeZone;
 
 //@UseReporter(FileLauncherReporter.class)
 public class VelocityUtilsTest
@@ -24,26 +24,15 @@ public class VelocityUtilsTest
   @Test
   public void testDate()
   {
-    try (ConsistentTimeZone tz = new ConsistentTimeZone()) {
+    /*
+    try (WithTimeZone tz = new WithTimeZone()) {
       Timestamp date = DateUtils.parse("2001/02/03");
       VelocityApprovals.verify(c -> {
         c.put("date", date);
       });
     }
+
+     */
   }
 
-  static class ConsistentTimeZone implements AutoCloseable {
-
-    private final TimeZone tz;
-
-    public ConsistentTimeZone() {
-        tz = TimeZone.getDefault();
-        TimeZone.setDefault(TimeZone.getTimeZone("UTC"));
-    }
-
-    @Override
-    public void close() {
-        TimeZone.setDefault(tz);
-    }
-  }
 }
